WIP: preliminary git-annex support
Supysonic doesn't deal well with [git-annex][] (GA)
repositories. Because GA stores files as symlinks to an object storage
in `.git/annex/objects`, all the normal `Artist/Album/File.mp3` files
are ignored and instead the actual files in `.git` are parsed.

 [git-annex]: https://git-annex.branchable.com/

Those files have funky filenames based on the file hash, for example:

```
.git/annex/objects/xm/zw/SHA256E-s1093632--6dc57e45ba7f81ed48ab6e581335bd2fb6271e40934c97931c6684cea3eec0b9.mp3/SHA256E-s1093632--6dc57e45ba7f81ed48ab6e581335bd2fb6271e40934c97931c6684cea3eec0b9.mp3
```

This makes most Supysonic client show the garbled filename instead of
the song title and breaks most album grouping behavior.

This patch makes the scanner follows symlinks when scanning and
ignores hidden directories (so `.git`). It's a WIP for now because
ideally the parser would be rewritten to use the faster `os.walk` and
support arbitrary ignore patterns or enable the user to configure
whether symlinks must be followed.

But I've been using this patch on my side and it works sufficiently
well.