Fix wrong mimetype being returned when transcoding
Previously, when a client that was configured to always transcode
requested a stream without specifying a format, the mimetype would not
be set to the transcode target type.

This commit simplifies the logic that decides which format should be
used and always sets the mimetype if the destination format doesn't
match the source format.