Fix bitrate units when scanning
In `0183bcb6` the scanner switched from using Mutagen to Mediafile for
scanning files. Prior to this commit, the bitrate from Mutagen was
divided by 1000 to convert it from bps to kbps. After switching to
Mediafile, the conversion was dropped even though Mediafile also reports
bitrate in bps.

This commit adds back the conversion to kbps and adds a test that checks
that the bitrate and some other metadata is correct.

This commit will fix transcoding being applied in some cases where it
isn't needed. This was happening because the bitrate in the DB was
always larger than the requested bitrate due to its units.